chore: remove images
The images in the root of the repository are outdated. The images in the
`./Images` folder are correct, but I would rather have a separate pull
request to bring these in since they should not be here anyway; they
should have been removed in the revert commit and I am not sure why they
were not.